+++ src/index.html
@@ -10,7 +10,6 @@
   <title>Tomodachi</title>
 </head>
 <body>
-
   <header>
     <nav>
       <ul>
@@ -26,8 +25,7 @@
       </ul>
     </nav>
   </header>
-  <main class="container">
-    
+  <main class="container">    
   </main>  
  
   <!--FIREBASE-->
@@ -46,6 +44,5 @@
 
      <!--Custom Code-->
      <script type="module" src="./scripts/login-out.js"></script>
-     <script type="module" src="./scripts/fs-login.js"></script>
 </body>
 </html>
