+++ src/login-out.css
@@ -6,9 +6,8 @@
 }
 
 header{
-    display:block;
+    display: none;
 }
-
 
 body,button,
 input {
@@ -217,19 +216,17 @@ form.sign-up-form{
     height: 41px;
     font-size: 1rem;
 }
-
+.image{
+    transition: 1.1s .4s ease-in-out;
+}
 .right-panel .content, .right-panel .image{
     transform: translateX(800px);
 }
-
 .bnh{
     width: 100%;
 } 
 .knj{
     width: 90%;
-}
-.image{
-    transition: 1.1s .4s ease-in-out;
 }
 
 /*ANIMACIÓN*/
