+++ src/router.js
@@ -1,2 +1,3 @@
 import { components } from './views/components.js';
 
+console.log(components);
