+++ src/views/components.js
@@ -1,13 +1,9 @@
 import Profile from './profile.js';
 import MainPage from './appMainPage.js';
 
-
-
 const components = {
-    profile: Profile,
-    mainPage: MainPage
+  profile: Profile,
+  mainPage: MainPage,
+};
 
- }
-
-
-export { components }
+export { components };
